Vote API is working

+++ backend/Models/Vote.cs
@@ -7,9 +7,9 @@ namespace Rate.ME.Models
     {
         public long Id { get; set; }
         public long TokenId { get; set; }
-        public long ProductRate { get; set; }
-        public long ServiceRate { get; set; }
-        public long RatioRate { get; set; }
+        public double ProductRate { get; set; }
+        public double ServiceRate { get; set; }
+        public double RatioRate { get; set; }
         public string Comment { get; set; }
 
         public virtual Token Token { get; set; }
